\name{sum-package}
\alias{sum-package}
\alias{sum}
\docType{package}
\title{
\packageTitle{sum}
}
\description{
\packageDescription{sum}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sum}
\packageIndices{sum}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sum}

Maintainer: \packageMaintainer{sum}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
